feat: "Correciones PR2"

+++ src/main/java/edu/dosw/dto/ReportDTO.java
@@ -1,14 +1,11 @@
 package edu.dosw.dto;
 
-/**
- * DTO for reporting reassignment statistics
- */
+/** DTO for reporting reassignment statistics */
 public record ReportDTO(
-        long total,
-        long pending,
-        long approved,
-        long rejected,
-        long cancellations,
-        long swaps,
-        long joins
-) {}
+    Integer total,
+    Integer pending,
+    Integer approved,
+    Integer rejected,
+    Integer cancellations,
+    Integer swaps,
+    Integer joins) {}
